Improve executable list check.
Executable requirements could be specified as either a list or a tuple.
When specified as a list, at least one item is expected to be installed.
When specified as a tuple, all items are expected to be installed.

This commit removes the assumption that only the first item of the
executable list should be installed and properly implement the check.